# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  price: number;
  imageUrl?: string | null;
  prepTimeLimit: number;
  isAvailable: boolean;
  createdAt?: string;
}

export interface CreateMenuItemRequest {
  name: string;
  category: string;
  price: number;
  imageUrl?: string | null;
  prepTimeLimit: number;
  isAvailable: boolean;
}

export interface MenuItemsResponse {
  items: MenuItem[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Event {
  id: number;
  name: string;
  discountPercentage: number;
  startDate: string;
  endDate: string;
  isActive: boolean;
}

export interface CreateEventRequest {
  name: string;
  discountPercentage: number;
  startDate: string;
  endDate: string;
  isActive: boolean;
}

export interface ActiveEventResponse {
  event: Event | null;
}

export type OrderItemDetails = { [key: string]: unknown };

export type OrderCustomerInfo = { [key: string]: unknown };

export interface Order {
  id: number;
  itemDetails: OrderItemDetails;
  totalPrice: number;
  customerInfo: OrderCustomerInfo;
  status: string;
  createdAt: string;
}

export type CreateOrderRequestItemDetails = { [key: string]: unknown };

export type CreateOrderRequestCustomerInfo = { [key: string]: unknown };

export interface CreateOrderRequest {
  itemDetails: CreateOrderRequestItemDetails;
  totalPrice: number;
  customerInfo: CreateOrderRequestCustomerInfo;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface OrdersResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Ad {
  id: number;
  imageUrl: string;
  title: string;
  subText?: string | null;
  link?: string | null;
  isActive: boolean;
}

export interface CreateAdRequest {
  imageUrl: string;
  title: string;
  subText?: string | null;
  link?: string | null;
  isActive: boolean;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AuthStatusResponse {
  authenticated: boolean;
  username?: string | null;
}

export type ListMenuItemsParams = {
  page?: number;
  limit?: number;
  category?: string;
  search?: string;
};

export type ListOrdersParams = {
  status?: string;
  page?: number;
  limit?: number;
};