# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu item operations
  - name: events
    description: Event and discount operations
  - name: orders
    description: Order operations
  - name: ads
    description: Advertisement operations
  - name: auth
    description: Authentication operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItemsResponse"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemRequest"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemRequest"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /menu/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List all menu categories
      responses:
        "200":
          description: List of category strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/active:
    get:
      operationId: getActiveEvent
      tags: [events]
      summary: Get the currently active event (if any)
      responses:
        "200":
          description: Active event or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveEventResponse"

  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List all ads
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Create a new ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdRequest"
      responses:
        "201":
          description: Created ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  /ads/active:
    get:
      operationId: listActiveAds
      tags: [ads]
      summary: List active ads for the hero banner
      responses:
        "200":
          description: List of active ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"

  /ads/{id}:
    put:
      operationId: updateAd
      tags: [ads]
      summary: Update an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdRequest"
      responses:
        "200":
          description: Updated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [ads]
      summary: Delete an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getAuthStatus
      tags: [auth]
      summary: Get current auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        prepTimeLimit:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - price
        - prepTimeLimit
        - isAvailable

    CreateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        prepTimeLimit:
          type: integer
        isAvailable:
          type: boolean
      required:
        - name
        - category
        - price
        - prepTimeLimit
        - isAvailable

    MenuItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - totalPages

    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        discountPercentage:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - discountPercentage
        - startDate
        - endDate
        - isActive

    CreateEventRequest:
      type: object
      properties:
        name:
          type: string
        discountPercentage:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - discountPercentage
        - startDate
        - endDate
        - isActive

    ActiveEventResponse:
      type: object
      properties:
        event:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Event"
      required:
        - event

    Order:
      type: object
      properties:
        id:
          type: integer
        itemDetails:
          type: object
          additionalProperties: true
        totalPrice:
          type: number
        customerInfo:
          type: object
          additionalProperties: true
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - itemDetails
        - totalPrice
        - customerInfo
        - status
        - createdAt

    CreateOrderRequest:
      type: object
      properties:
        itemDetails:
          type: object
          additionalProperties: true
        totalPrice:
          type: number
        customerInfo:
          type: object
          additionalProperties: true
      required:
        - itemDetails
        - totalPrice
        - customerInfo

    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
      required:
        - status

    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages

    Ad:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        title:
          type: string
        subText:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - imageUrl
        - title
        - isActive

    CreateAdRequest:
      type: object
      properties:
        imageUrl:
          type: string
        title:
          type: string
        subText:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - imageUrl
        - title
        - isActive

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    AuthStatusResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: string
          nullable: true
      required:
        - authenticated