# lib/api-zod/src/generated/types/createMenuItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemRequest {
  name: string;
  category: string;
  price: number;
  imageUrl?: string | null;
  prepTimeLimit: number;
  isAvailable: boolean;
}