# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestCustomerInfo } from "./createOrderRequestCustomerInfo";
import type { CreateOrderRequestItemDetails } from "./createOrderRequestItemDetails";

export interface CreateOrderRequest {
  itemDetails: CreateOrderRequestItemDetails;
  totalPrice: number;
  customerInfo: CreateOrderRequestCustomerInfo;
}