# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCustomerInfo } from "./orderCustomerInfo";
import type { OrderItemDetails } from "./orderItemDetails";

export interface Order {
  id: number;
  itemDetails: OrderItemDetails;
  totalPrice: number;
  customerInfo: OrderCustomerInfo;
  status: string;
  createdAt: string;
}